extract get_editable_role_choices
